pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jmx-to-scenario"
version = "1.0.0"
description = "Convert JMeter JMX files to pt_scenario.yaml format"
readme = "SPEC.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "defusedxml>=0.7.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]

[project.scripts]
jmx-to-scenario = "jmx_to_scenario.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/jmx_to_scenario"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
